moved external SBOL path as a function to SBOLUtility.
